feat(ftplugin/quarto_rnvim.lua): prevent multiple setup calls
This commit prevents multiple calls to the setup functions in the
`quarto_rnvim.lua` file. This is achieved by introducing a buffer-local
variable `vim.b.rnvim_quarto_config`. The setup functions are only called
if this variable is not already set. This prevents issues that may arise
from running the setup functions multiple times, such as duplicate
mappings or incorrect configurations.